% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListPathway.R
\name{ListPathway}
\alias{ListPathway}
\title{List available pathways or pathway metadata}
\usage{
ListPathway(query = NULL, drop_empty = TRUE)
}
\arguments{
\item{query}{Optional character.
\itemize{
\item NULL: return full summary table
\item "Pathway": return unique pathway names
\item pathway name (e.g. "WNT", "NOTCH"): return rows for that pathway
}}

\item{drop_empty}{Logical; remove entries with 0 genes.
Default TRUE.}
}
\value{
A tibble or a character vector depending on query.
}
\description{
Reads the "Summary" sheet from Pathway_Database_Combined.xlsx.
Can return the full table, unique pathway names, or
rows for a specific pathway.
}
\examples{
ListPathway()
ListPathway("Pathway")
ListPathway("WNT")

}
