% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayDataSet.R
\name{load_pathway_data}
\alias{load_pathway_data}
\title{This function reads pathway data from the package's built-in Excel file.}
\usage{
load_pathway_data(pathway)
}
\arguments{
\item{pathway}{The name of the pathway interested.}
}
\value{
A data frame with pathway data.
}
\description{
This function reads pathway data from the package's built-in Excel file.
}
