% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pathwayembed.R
\name{pathwayembed}
\alias{pathwayembed}
\title{A function for scRNA sequencing pathway analysis}
\arguments{
\item{x}{A Seurat Object.}

\item{pathway}{A character.}

\item{idents}{Idents of the X}
}
\value{
The sum of x and y.
}
\description{
A function for scRNA sequencing pathway analysis
}
\examples{
pathwayembed(x, pathway, idents)
}
