% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayMaxMin.R
\name{PathwayMaxMin}
\alias{PathwayMaxMin}
\title{PathwayMaxMin}
\usage{
PathwayMaxMin(expr_data, pathwaydata)
}
\arguments{
\item{expr_data}{Pre-processed gene-by-cell matrix (z-scored, pathway-filtered)}

\item{pathwaydata}{Pathway data outcome from LoadPathway()}
}
\value{
A data.frame with pathway.on and pathway.off values per gene
}
\description{
A function to obtain the hypothetical max and min activation status of selected pathway for a given scRNA seq data set
}
\examples{
data(fake_test_matrix)
PathwayMaxMin(fake_test_matrix, pathwaydata)
}
