% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreparePlotData.R
\name{PreparePlotData}
\alias{PreparePlotData}
\title{A function to prepare the signal transduction dataframe for plotting}
\usage{
PreparePlotData(x, final_mds, group)
}
\arguments{
\item{x}{A\code{matrix} framework containing cell ID x metadata.}

\item{final_mds}{A 'dataframe' output from ComputeCellData.}

\item{group}{group for the comparision}
}
\value{
data for plotting
}
\description{
A function to prepare the signal transduction dataframe for plotting
}
\examples{
data(fake_test_metadata)
data(fake_final_mds)
PreparePlotData(fake_test_metadata, fake_final_mds, "genotype")
}
