% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculatePercentage.R
\name{CalculatePercentage}
\alias{CalculatePercentage}
\title{Calculate the percentage of cells in activation status}
\usage{
CalculatePercentage(to.plot, group_var)
}
\arguments{
\item{to.plot}{A data frame containing at least a \code{scale} column and a grouping column.}

\item{group_var}{A string specifying the grouping variable (e.g., "genotype", "treatment").}
}
\value{
A data frame with the percentage of ON/OFF cells and Cohen's d (if applicable).
}
\description{
Calculate the percentage of cells in activation status
}
\examples{
data(fake_to_plot)
CalculatePercentage(fake_to_plot, "genotype")
}
