% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeCellData.R
\name{ComputeCellData}
\alias{ComputeCellData}
\title{ComputeCellData}
\usage{
ComputeCellData(expr_data, pathway.stat, distance.method = "manhattan")
}
\arguments{
\item{expr_data}{Pre-processed gene-by-cell matrix (z-scored)}

\item{pathway.stat}{Data.frame from PathwayMaxMin}

\item{distance.method}{Distance metric (default = "manhattan")}
}
\value{
A named numeric vector of cell scores between 0 (OFF) and 1 (ON)
}
\description{
Compute cell status for a given pathway by measuring distances
between cells and pathway ON/OFF reference states.
}
