% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeCellData.R
\name{ComputeCellData}
\alias{ComputeCellData}
\title{A function for scRNA sequencing pathway analysis}
\usage{
ComputeCellData(x, pathway, distance.method)
}
\arguments{
\item{x}{A \code{Seurat} object containing single-cell RNA sequencing data.}

\item{pathway}{A \code{character} string specifying the pathway name.}

\item{distance.method}{A \code{character} string specifying the distance method.
Options include: "manhattan", "euclidean", "canberra", "binary", "minkowski".}
}
\value{
A data frame representing the multidimensional scaling (MDS) results
for the cells based on the pathway expression.
}
\description{
This function computes cell status for a given pathway in single-cell RNA-seq data,
based on the distance between genes in a specified pathway. The distance is computed
for each batch of cells, and classical multidimensional scaling (MDS) is used to
visualize the pathway expression across cells.
}
\examples{
data(fake_test_object) # load the fake test data
ComputeCellData(fake_test_object, "Wnt", "manhattan")
}
