% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeCellData.R
\name{ComputeCellData}
\alias{ComputeCellData}
\title{ComputeCellData}
\usage{
ComputeCellData(
  x,
  pathway,
  distance.method,
  batch.size = batch.size,
  scale.data = TRUE
)
}
\arguments{
\item{x}{A \code{Seurat} object containing single-cell RNA sequencing data.}

\item{pathway}{A \code{character} string specifying the pathway name. This should match a pathway used by \code{LoadPathway()}.}

\item{distance.method}{A \code{character} string specifying the distance metric to use.Default is "manhattan".
Options include: \code{"manhattan"}, \code{"euclidean"}, \code{"canberra"}, \code{"binary"}, \code{"minkowski"}}

\item{batch.size}{An \code{integer} specifying the number of cells to process per batch. Default is 1000.}

\item{scale.data}{A \code{logical} indicating whether to use scaled data (\code{scale.data = TRUE}) or normalized data. Default is \code{TRUE}.}
}
\value{
A data frame of MDS results with normalized values per cell, suitable for thresholding or visualization.
}
\description{
A function computes cell status for a given pathway in single-cell RNA-seq data,
based on the distance between genes in a specified pathway. The distance is computed
for each batch of cells, and classical multidimensional scaling (MDS) is used to
visualize the pathway expression across cells.
}
\examples{
data(fake_test_object)
ComputeCellData(fake_test_object, pathway = "Wnt", distance.method = "manhattan", batch.size = 2000)

}
