% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPreProcess.R
\name{DataPreProcess}
\alias{DataPreProcess}
\title{DataPreProcess}
\usage{
DataPreProcess(
  x,
  pathwaydata,
  Seurat.object = FALSE,
  assay = "RNA",
  slot = "data"
)
}
\arguments{
\item{x}{A Seurat object OR a gene-by-cell normalized expression matrix}

\item{pathwaydata}{Pathway data from the output of LoadPathway()}

\item{Seurat.object}{Logical; whether x is a Seurat object}

\item{assay}{Assay to use if x is a Seurat object (default = "RNA")}

\item{slot}{Slot to extract from Seurat object (default = "data")}
}
\value{
A z-scored expression matrix of pathway genes x cell
}
\description{
Preprocess expression data for pathway analysis.
}
\examples{
DataPreProcess(seurat_obj, pathwaydata, Seurat.object = TRUE)
DataPreProcess(norm_matrix, pathwaydata)

}
