% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{fake_final_mds}
\alias{fake_final_mds}
\title{Example Cell Status and Normalized Data}
\format{
A data frame with the following columns:
\describe{
\item{V1}{A numerical value presenting the status of the cell}
\item{normalized}{Numerical value representing normalized data}
}
}
\usage{
data(fake_final_mds)
}
\description{
A dataset generated by \code{ComputeCellData} applied to \code{fake_test_object}.
Contains cell status and normalized values.
}
\examples{
data(fake_final_mds)
head(fake_final_mds)
}
\keyword{datasets}
