% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{fake_to_plot}
\alias{fake_to_plot}
\title{Example Processed Data for Plotting}
\format{
A data frame with the following columns:
\describe{
\item{V1}{A numerical or categorical value depending on the specific analysis}
\item{normalized}{Normalized numerical value representing the cell's data}
\item{genotype}{Group classification of the cell (e.g., "WT", "Mutant")}
\item{scale}{Scaled data for visualization purposes}
}
}
\usage{
data(fake_to_plot)
}
\description{
A dataset generated by running \code{PreparePlotData} on \code{fake_test_object} and \code{fake_final_mds}.
Contains processed data ready for visualization, with the following features:
}
\examples{
data(fake_to_plot)
head(fake_to_plot)
}
\keyword{datasets}
