% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadPathway.R
\name{LoadPathway}
\alias{LoadPathway}
\title{This function reads pathway data from the package's built-in Excel file.}
\usage{
LoadPathway(pathway)
}
\arguments{
\item{pathway}{A \code{character} string specifying the pathway name.}
}
\value{
A data frame with pathway data.
}
\description{
This function reads pathway data from the package's built-in Excel file.
}
\examples{
LoadPathway("Wnt")
}
