% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayMaxMin.R
\name{PathwayMaxMin}
\alias{PathwayMaxMin}
\title{PathwayMaxMin}
\usage{
PathwayMaxMin(x, pathway, scale.data = TRUE)
}
\arguments{
\item{x}{A \code{matrix} of genes x cells}

\item{pathway}{A \code{character} string specifying the pathway name.}

\item{scale.data}{A \code{logical} indicating whether to use scaled data (\code{scale.data = TRUE}) or normalized data. Default is \code{TRUE}.}
}
\value{
The hypothetical value for Pathway on and off (max and min value for features)
}
\description{
A function to obtain the hypothetical max and min activation status of selected pathway for a given scRNA seq data set
}
\examples{
data(fake_test_matrix) # load the fake test data
PathwayMaxMin(fake_test_matrix, "Wnt", scale.data = TRUE)
}
