% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayMaxMin.R
\name{PathwayMaxMin}
\alias{PathwayMaxMin}
\title{A function for scRNA sequencing pathway analysis}
\usage{
PathwayMaxMin(x, pathway)
}
\arguments{
\item{x}{A Seurat Object.}

\item{pathway}{The name of the pathway.}
}
\value{
The value for Pathway on and off (max and min value for features)
}
\description{
A function for scRNA sequencing pathway analysis
}
\examples{
data(fake_test_object) # load the fake test data
PathwayMaxMin(fake_test_object, "Wnt")
}
