% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPathway.R
\name{PlotPathway}
\alias{PlotPathway}
\title{A function to plot the Pathway activation status}
\usage{
PlotPathway(to.plot, pathway, group, color)
}
\arguments{
\item{to.plot}{A data frame with pathway activation values genereated by PreparePlotData}

\item{pathway}{A character string indicating the pathway name.}

\item{group}{Column name to group and color by (e.g., genotype).}

\item{color}{A character vector of colors to use for fill and outline.}
}
\value{
A ggplot object.
}
\description{
A function to plot the Pathway activation status
}
\examples{
data(fake_to_plot)
PlotPathway(to.plot = fake_to_plot,"Wnt","genotype",color = c("#ae282c", "#2066a8"))
}
