% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPathway.R
\name{PlotPathway}
\alias{PlotPathway}
\title{A function to plot the Pathway activation status}
\usage{
PlotPathway(to.plot, pathway, group, color)
}
\arguments{
\item{to.plot}{A dataframe.}

\item{pathway}{A name of the pathway.}

\item{group}{Ident of the plot.}

\item{color}{Colors for the group.}
}
\value{
A plot.
}
\description{
A function to plot the Pathway activation status
}
\examples{
data(fake_to_plot)
PlotPathway(fake_to_plot, "Wnt", "genotype", c("#ae282c","#2066a8"))
}
